서버 validation check 기능추가

+++ src/main/java/com/example/util/PageMarker.java
@@ -4,12 +4,12 @@ package com.example.util;
 public class PageMarker {
 
     private Criteria cri;
-    private int totalCount;
-    private int startPage;
-    private int endPage;
+    private long totalCount;
+    private long startPage;
+    private long endPage;
     private boolean prev;
     private boolean next;
-    private int displayPageNum = 5;
+    private long displayPageNum = 5;
 
     public Criteria getCri() {
         return cri;
@@ -17,22 +17,22 @@ public class PageMarker {
     public void setCri(Criteria cri) {
         this.cri = cri;
     }
-    public int getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
-    public void setTotalCount(int totalCount) {
+    public void setTotalCount(long totalCount) {
         this.totalCount = totalCount;
         calcData();
     }
 
     private void calcData() {
 
-        endPage = (int) (Math.ceil(cri.getPage() / (double) displayPageNum) * displayPageNum);
+        endPage = (long) (Math.ceil(cri.getPage() / (double) displayPageNum) * displayPageNum);
 
         startPage = (endPage - displayPageNum) + 1;
         if(startPage <= 0) startPage = 1;
 
-        int tempEndPage = (int) (Math.ceil(totalCount / (double) cri.getPerPageNum()));
+        long tempEndPage = (long) (Math.ceil(totalCount / (double) cri.getPerPageNum()));
         if (endPage > tempEndPage) {
             endPage = tempEndPage;
         }
@@ -42,16 +42,16 @@ public class PageMarker {
 
     }
 
-    public int getStartPage() {
+    public long getStartPage() {
         return startPage;
     }
-    public void setStartPage(int startPage) {
+    public void setStartPage(long startPage) {
         this.startPage = startPage;
     }
-    public int getEndPage() {
+    public long getEndPage() {
         return endPage;
     }
-    public void setEndPage(int endPage) {
+    public void setEndPage(long endPage) {
         this.endPage = endPage;
     }
     public boolean isPrev() {
@@ -66,10 +66,10 @@ public class PageMarker {
     public void setNext(boolean next) {
         this.next = next;
     }
-    public int getDisplayPageNum() {
+    public long getDisplayPageNum() {
         return displayPageNum;
     }
-    public void setDisplayPageNum(int displayPageNum) {
+    public void setDisplayPageNum(long displayPageNum) {
         this.displayPageNum = displayPageNum;
     }
